pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "gunicorn>=25.1.0",
    "rollbar>=1.3.0",
    "whitenoise>=6.12.0",
]
[tool.ruff.lint]
# "E" и "W" включают все ошибки и предупреждения стиля (включая пустые строки)
select = ["E", "F", "W", "I"]
# Обязательно убери игнорирование E302, если оно там было
ignore = []

[tool.ruff]
# Убедись, что длина строки совпадает с требованиями Хекслета (обычно 79)
line-length = 79